+++ CHANGELOG.rst
@@ -4,7 +4,9 @@ Changelog
 0.8.0 (unreleased)
 ------------------
 
-- Nothing changed yet.
+**Bug fixes**
+
+- Fix crash when the file is not uploaded using ``attachment`` field name (fixes #57)
 
 
 0.7.0 (2016-06-10)