+++ CHANGELOG.rst
@@ -8,6 +8,9 @@ Changelog
 
 - Fix crash when the file is not uploaded using ``attachment`` field name (fixes #57)
 
+**Internal changes**
+
+- Remove hard-coded CORS setup (fixes #59)
 
 0.7.0 (2016-06-10)
 ------------------