+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Changelog
 **Bug fixes**
 
 - Fix crash when the file is not uploaded using ``attachment`` field name (fixes #57)
+- Prevent crash when filename is not provided (fixes #81)
 
 
 0.7.0 (2016-06-10)