+++ CHANGELOG.rst
@@ -12,10 +12,12 @@ Changelog
 
 - Fix crash when the file is not uploaded using ``attachment`` field name (fixes #57)
 - Prevent crash when filename is not provided (fixes #81)
+- Update the call to the Record resource to use named attributes. (#97)
 
 **Internal changes**
 
 - Remove hard-coded CORS setup (fixes #59)
+
 
 0.7.0 (2016-06-10)
 ------------------